Finalize rename from ButtonNew to Button
