+++ src/app/backing/components/Spotlight/Spotlight.tsx
@@ -4,7 +4,7 @@ import { useHandleErrors } from '@/app/collective-rewards/utils'
 import { BackMoreBuildersCard, BuilderCardControl } from '@/app/shared/components/BuilderCard'
 import { BuildersSpotlight } from '@/app/shared/components/BuildersSpotlight'
 import { useGetBuilderEstimatedRewards } from '@/app/shared/hooks/useGetBuilderEstimatedRewards'
-import { Button } from '@/components/ButtonNew'
+import { Button } from '@/components/Button'
 import { LoadingSpinner } from '@/components/LoadingSpinner'
 import { useRouter, useSearchParams } from 'next/navigation'
 import { useCallback, useContext, useEffect, useMemo } from 'react'
