+++ src/app/builders/components/BecomeBuilderBanner/BecomeBuilderBanner.tsx
@@ -1,6 +1,6 @@
 import { Paragraph } from '@/components/TypographyNew'
 import { Typography } from '@/components/TypographyNew/Typography'
-import { Button } from '@/components/ButtonNew'
+import { Button } from '@/components/Button'
 import Image from 'next/image'
 import CollapsibleWithPreview from '@/components/CollapsibleWithPreview/CollapsibleWithPreview'
 import { useRouter } from 'next/navigation'
