+++ src/app/communities/nft/[address]/_components/ClaimItButton.tsx
@@ -1,5 +1,5 @@
 'use client'
-import { Button } from '@/components/ButtonNew'
+import { Button } from '@/components/Button'
 import { Span, Paragraph } from '@/components/TypographyNew'
 import { useCommunityNFT } from '@/app/communities/nft/[address]/CommunityNFTContext'
 import { useAccount } from 'wagmi'
