+++ src/app/communities/nft/[address]/_sections/CommunityInfoHeader.tsx
@@ -9,7 +9,7 @@ import { communitiesMapByContract } from '@/app/communities/communityUtils'
 import { cn, truncateMiddle } from '@/lib/utils'
 import { ImageMask } from '@/components/ImageMask'
 import { AddToWalletButton } from '../_components/AddToWalletButton'
-import { Button } from '@/components/ButtonNew'
+import { Button } from '@/components/Button'
 import { Header, Paragraph } from '@/components/TypographyNew'
 import { CopyButton } from '@/components/CopyButton'
 import { useNFTBoosterContext } from '@/app/providers/NFT/BoosterContext'
