+++ src/app/my-rewards/backers/components/Table/BackerRewardsHeaderRow.tsx
@@ -9,7 +9,7 @@ import {
   TableColumnDropdown,
 } from '@/app/builders/components/Table'
 import { Action } from '@/app/builders/components/Table/Cell/ActionCell'
-import { Button } from '@/components/ButtonNew'
+import { Button } from '@/components/Button'
 import { CloseIconKoto } from '@/components/Icons'
 import { TableHeaderCell } from '@/components/TableNew'
 import { Paragraph } from '@/components/TypographyNew'
