+++ src/app/proposals/new/components/NewProposalCardExtended.stories.tsx
@@ -3,7 +3,7 @@ import { NewProposalCardExtended } from './NewProposalCardExtended'
 import { newProposalCards, type NewProposalCardExtendedData } from '../newProposalCards.data'
 import { useState } from 'react'
 import { AnimatePresence } from 'motion/react'
-import { Button } from '@/components/ButtonNew'
+import { Button } from '@/components/Button'
 
 const meta: Meta<typeof NewProposalCardExtended> = {
   title: 'Proposals/NewProposalCardExtended',
