+++ src/app/proposals/page.tsx
@@ -4,7 +4,7 @@ import { withFallbackRetry } from '@/app/shared/components/Fallback/FallbackWith
 import { ProposalsFromChain } from '@/app/proposals/ProposalsFromChain'
 import { ProposalsFromTheGraph } from '@/app/proposals/ProposalsFromTheGraph'
 import { HeroComponent } from '@/components/HeroComponent'
-import { Button } from '@/components/ButtonNew'
+import { Button } from '@/components/Button'
 import { useRouter } from 'next/navigation'
 import { useAccount } from 'wagmi'
 import { ConnectWorkflow } from '@/shared/walletConnection/connection/ConnectWorkflow'
