+++ src/app/shared/components/BuilderCard/BuilderCardControl.tsx
@@ -2,7 +2,7 @@ import { AllocationsContext } from '@/app/collective-rewards/allocations/context
 import { useAllocateVotes } from '@/app/collective-rewards/allocations/hooks/useAllocateVotes'
 import { Builder } from '@/app/collective-rewards/types'
 import { TransactionInProgressButton } from '@/app/user/Stake/components/TransactionInProgressButton'
-import { Button } from '@/components/ButtonNew/Button'
+import { Button } from '@/components/Button/Button'
 import { useLayoutContext } from '@/components/MainContainer/LayoutProvider'
 import { ActionsContainer } from '@/components/containers/ActionsContainer'
 import { RIF } from '@/lib/constants'
