+++ src/app/user/Balances/BalanceInfoForUser.tsx
@@ -4,7 +4,7 @@ import { BalanceInfo } from '@/components/BalanceInfo'
 import Big from '@/lib/big'
 import { formatCurrency, formatCurrencyWithLabel } from '@/lib/utils'
 import { Paragraph } from '@/components/TypographyNew'
-import { Button } from '@/components/ButtonNew'
+import { Button } from '@/components/Button'
 import { useState } from 'react'
 import { requestProviderToAddToken } from '@/shared/utils'
 import { RBTC } from '@/lib/constants'
