+++ src/app/user/Balances/BalancesSection.tsx
@@ -5,7 +5,7 @@ import { useModal } from '@/shared/hooks/useModal'
 import { useEffect, useState } from 'react'
 import { useSearchParams } from 'next/navigation'
 import { useBalancesContext } from '@/app/user/Balances/context/BalancesContext'
-import { Button } from '@/components/ButtonNew'
+import { Button } from '@/components/Button'
 import { MoneyIconKoto } from '@/components/Icons'
 import { useRef } from 'react'
 import { Span } from '@/components/TypographyNew'
