+++ src/app/user/Stake/components/StepActionButtons.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/components/ButtonNew/Button'
+import { Button } from '@/components/Button/Button'
 import { cn } from '@/lib/utils'
 import { ReactNode } from 'react'
 import { TransactionInProgressButton } from './TransactionInProgressButton'
