+++ src/app/user/Unstake/components/PercentageButtons.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/components/ButtonNew/Button'
+import { Button } from '@/components/Button/Button'
 import { Span } from '@/components/TypographyNew'
 
 interface Props {
