+++ src/app/user/Unstake/components/UnstakeActions.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/components/ButtonNew/Button'
+import { Button } from '@/components/Button/Button'
 import { Divider } from '@/components/Divider'
 import { Hash } from 'viem'
 import { TransactionInProgressButton } from '@/app/user/Stake/components/TransactionInProgressButton'
