+++ src/components/TableNew/TablePager.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/components/ButtonNew/Button'
+import { Button } from '@/components/Button/Button'
 import { Paragraph, Span } from '@/components/TypographyNew'
 import { CommonComponentProps } from '@/components/commonProps'
 import { cn } from '@/lib/utils'
