+++ src/components/Tooltip/Tooltip.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import type { Meta, StoryObj } from '@storybook/nextjs'
 import { Tooltip } from './Tooltip'
-import { Button } from '../ButtonNew/Button'
+import { Button } from '../Button/Button'
 import { Paragraph } from '../TypographyNew'
 
 const meta: Meta<typeof Tooltip> = {
