+++ src/shared/walletConnection/components/DisconnectButton.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/components/ButtonNew/Button'
+import { Button } from '@/components/Button/Button'
 import { PowerIconKoto } from '@/components/Icons'
 
 interface Props {
